Update Blood2.asl
Changed to make it split on level transitions instead of episode endings. Added option for episode only splitting if people only want 4 splits.

Changed it so you can start on any episode and do the episodes in any order after that as well, no more being locked to specific orders.